pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ali-gater"
version = "0.1.0"
description = "ali-gater: A lightweight principle driven checker."
authors = [
    {name = "Sergio M. Ferro", email = "Sergio_Ferro@epam.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "ragas>=0.0.22",
    "unstructured[pdf]>=0.11.8",
    "python-dotenv>=1.0.0",
    "pytest>=8.4.2",
    "beautifulsoup4>=4.14.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ali-gater"
Documentation = "https://github.com/yourusername/ali-gater#readme"
Repository = "https://github.com/yourusername/ali-gater"
Issues = "https://github.com/yourusername/ali-gater/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ali_gater"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=ali_gater --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
